+++ backend/src/config/api-config.module.ts
@@ -1,6 +1,7 @@
-import { Module } from '@nestjs/common';
-import { ApiConfigService } from './ApiConfig.service';
 import { ConfigService } from '@nestjs/config';
+import { Module } from '@nestjs/common';
+
+import { ApiConfigService } from './api-config.service';
 
 @Module({
   providers: [ApiConfigService, ConfigService],
